4.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,8 +1,7 @@
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
-// const User = require('../models/user');
-// const jwt = require('jsonwebtoken');
-const middleware = require('../utils/middleware');
+const User = require('../models/user');
+const jwt = require('jsonwebtoken');
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', {
@@ -22,10 +21,15 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 });
 
-blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+blogsRouter.post('/', async (request, response) => {
   const body = request.body;
 
-  const user = request.user;
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+
+  const user = await User.findById(decodedToken.id);
 
   const blog = new Blog({
     title: body.title,
@@ -42,20 +46,21 @@ blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   response.json(savedBlog);
 });
 
-blogsRouter.delete(
-  '/:id',
-  middleware.userExtractor,
-  async (request, response) => {
-    const blogToDelete = await Blog.findById(request.params.id);
+blogsRouter.delete('/:id', async (request, response) => {
+  const blogToDelete = await Blog.findById(request.params.id);
 
-    const user = request.user;
-
-    if (blogToDelete.user.toString() === user.id.toString()) {
-      await Blog.findByIdAndDelete(blogToDelete.id);
-      response.status(204).end();
-    }
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' });
   }
-);
+
+  const user = await User.findById(decodedToken.id);
+
+  if (blogToDelete.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndDelete(blogToDelete.id);
+    response.status(204).end();
+  }
+});
 
 blogsRouter.put('/:id', async (request, response) => {
   const body = request.body;
